fix: restructuring of the project
