+++ app/[lng]/(auth)/(components)/AuthButton.js
