+++ app/[lng]/(auth)/(components)/Checkbox.js
