+++ app/[lng]/(auth)/forgot-password/page.js
