+++ app/[lng]/(auth)/register/page.js
