+++ app/[lng]/image/[title]-[uuid]/page.js
@@ -3,7 +3,7 @@
 const ImagePage = () => {
   return (
     <div>
-      <h1>Image Page</h1>
+      <h1 className="text-2xl font-bold text-gray-800">Image Page</h1>
     </div>
   )
 }
