+++ app/[lng]/pages/license/page.js
@@ -1,3 +1,3 @@
 export default function License() {
-  return <div>License Page</div>
+  return <div className="text-red-500">License Page</div>
 }
