+++ app/[lng]/pages/submition/page.js
@@ -1,3 +1,3 @@
 export default function Submition() {
-  return <div>Submition Page</div>
+  return <div className="text-red-500">Submition Page</div>
 }
