+++ app/[lng]/pages/toc/page.js
@@ -1,3 +1,3 @@
 export default function Toc() {
-  return <div>Terms of content page</div>
+  return <div className="text-red-500">Terms of content page</div>
 }
