+++ app/[lng]/pages/tou/page.js
@@ -1,3 +1,3 @@
 export default function Toc() {
-  return <div>Terms of users page</div>
+  return <div className="text-red-500">Terms of users page</div>
 }
